+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 
 - Added mining particles to Glowing Mushrooms (#32)
+- Added an alias to ModBlockEntityTypes for `redstone_interface` -> `basic_redstone_interface` to prevent data loss, so 1.5.0 is no longer a breaking change
 
 # 1.5.0